requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
pycparser==2.22
PyMySQL==1.1.1
qrcode==8.2
Werkzeug==3.1.3
mysql-connector-python
python-dotenv
gunicorn==20.1.0